requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
python-telegram-bot>=20.7
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0